std: move `Once` implementations to `sys`
