+++ library/std/src/sys/sync/once/no_threads.rs
