+++ library/std/src/sys/sync/once/queue.rs
