+++ library/std/src/sys/sync/rwlock/no_threads.rs
