+++ library/std/src/sys/sync/rwlock/solid.rs
